pyproject.toml
[project]
name = "todo_app"
version = "0.1.0"
description = "Evolution of Todo - Phase I"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "pydantic-settings",
    "python-dotenv",
    "python-jose[cryptography]",
    "bcrypt",
    "email-validator",
    "python-multipart",
    "openai",
    "mcp",
    "asyncpg"
]
requires-python = ">=3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
